+++ resources/js/addon.js
@@ -1,86 +1,55 @@
-import Span from './marks/Span'
-import ToolbarButton from "./components/ToolbarButton.vue";
-const { toggleBlockType } = Statamic.$bard.tiptap.commands;
+import Span from './marks/span'
+import Core from './extensions/core'
+import { styleToIcon } from './icons';
 
-const types = ['heading', 'paragraph', 'span'];
-
-const chars = {
-    heading: 'H',
-    paragraph: 'P',
-    span: 'T',
-};
-
-const tags = {
-    heading: 'h',
-    paragraph: 'p',
-    span: 'span',
-};
-
-const exts = {
-    heading: 'heading',
-    paragraph: 'paragraph',
-    span: 'bts_span',
+const types = {
+    heading: {
+        tag: 'h',
+        ext: 'heading',
+        cmd: 'btsToggleHeading'
+    },
+    paragraph: {
+        tag: 'p',
+        ext: 'paragraph',
+        cmd: 'btsToggleParagraph'
+    },
+    span: {
+        tag: 'span',
+        ext: 'bts_span',
+        cmd: 'btsToggleSpan'
+    },
 };
 
 Statamic.booting(() => {
 
-    const { mutator } = BardMutator;
+    // Initialization
 
-    Statamic.$bard.addExtension(() => new Span());
+    const { store, attr, styles, styleTypes } = Statamic.$config.get('bard-texstyle');
 
-    const store = Statamic.$config.get('statamic-bard-texstyle.store') || 'class';
-    const attr  = store === 'class' ? 'class' : 'bts_key';
+    // Extensions
 
-    const styles = Statamic.$config.get('statamic-bard-texstyle.styles') || [];
-    const activeTypes = _.uniq(Object.entries(styles).map(([, style]) => style.type));
-
-    const schemaMutator = (schema, { extendSchema }) => extendSchema(schema, {
-        attrs: {
-            [attr]: {
-                default: null,
-            },
-        },
-        parseDOMAttrs: dom => ({
-            [attr]: dom.getAttribute(`data-bard-${attr}`),
-        }),
-        toDOMAttrs: node => ({
-            [`data-bard-${attr}`]: node.attrs[attr],
-        }),
-    });
-
-    if (activeTypes.includes('heading')) {
-        mutator.schema('heading', schemaMutator);
-    }
-    if (activeTypes.includes('paragraph')) {
-        mutator.schema('paragraph', schemaMutator);
-        mutator.commands('paragraph', (commands, { type, schema }) => ({
-            ...commands,
-            [type.name]: attrs => toggleBlockType(type, schema.nodes.paragraph, attrs),
-        }));
-    }
-    if (activeTypes.includes('span')) {
-        mutator.schema('bts_span', schemaMutator);
-    }
+    Statamic.$bard.addExtension(() => Span);
+    Statamic.$bard.addExtension(() => Core.configure({ attr, styleTypes }));
 
     // Buttons
     
     Statamic.$bard.buttons((buttons, button) => {
         Object.entries(styles).forEach(([key, style]) => {
-            if (!types.includes(style.type)) {
+            if (!types[style.type]) {
                 return;
             }
-            const icon = [chars[style.type], style.ident || ''];
+            const type = types[style.type];
+            const icon = styleToIcon(style, type);
+            const args = style.type === 'heading'
+                ? { [attr]: style[store], level: style.level }
+                : { [attr]: style[store] };
             const data = {
                 name: key,
                 text: style.name,
-                command: exts[style.type],
-                args: style.type === 'heading'
-                    ? { [attr]: style[store], level: style.level }
-                    : { [attr]: style[store] },
-                component: ToolbarButton,
-                html: `<div class="bts-button"><span class="bts-button-char">${icon[0]}</span><sup class="bts-button-ident">${icon[1]}</sup></div>`,
-                bts_config: { store, attr },
-                bts_style: { ...style, icon },
+                args: args,
+                activeName: type.ext,
+                html: icon,
+                command: (editor, args) => editor.chain().focus()[type.cmd](args).run(),
             };
             buttons.splice(buttons.indexOf(key), 0, button(data));
         });
@@ -90,13 +59,14 @@ Statamic.booting(() => {
 
     const css = [];
     Object.entries(styles).forEach(([, style]) => {
-        if (!types.includes(style.type)) {
+        if (!types[style.type]) {
             return;
         }
+        const type = types[style.type];
         const tag = style.type === 'heading'
-            ? `${tags[style.type]}${style.level}`
-            : `${tags[style.type]}`;
-        css.push(`.bard-fieldtype .ProseMirror ${tag}[data-bard-${attr}="${style[store]}"] { ${style.cp_css} }`);
+            ? `${type.tag}${style.level}`
+            : `${type.tag}`;
+        css.push(`.bard-fieldtype .ProseMirror ${tag}[data-bts="${style[store]}"] { ${style.cp_css} }`);
     });
     const el = document.createElement('style');
     el.appendChild(document.createTextNode(css.join(' ')));