+++ src/Marks/Span.php
@@ -2,13 +2,23 @@
 
 namespace JackSleight\StatamicBardTexstyle\Marks;
 
-use JackSleight\StatamicBardMutator\Support\Traits\MutatesMark;
-use ProseMirrorToHtml\Marks\Mark;
+use Tiptap\Core\Mark;
 
 class Span extends Mark
 {
-    use MutatesMark;
+    public static $name = 'bts_span';
 
-    protected $markType = 'bts_span';
-    protected $tagName = 'span';
+    public function parseHTML()
+    {
+        return [
+            [
+                'tag' => 'span',
+            ],
+        ];
+    }
+
+    public function renderHTML($mark, $HTMLAttributes = [])
+    {
+        return ['span', $HTMLAttributes, 0];
+    }
 }